CI, better colors and logging
