+++ TODO.md
@@ -1,9 +1,6 @@
 
-tests -> something is done
-
 tab completion -> waiting for PR on Sbt
 
-CI
+firing a good error when sbt has not the right version --> not sure how to reliably do that
 
-firing a good error when sbt has not the right version
 check timeouts when sbt have to be downloaded...
