+++ build.sbt
@@ -5,7 +5,7 @@ name := "sbtcli"
 
 organization := "org.akka-js"
 
-scalaVersion := "2.12.6"
+scalaVersion := "2.12.7"
 
 fork in run := true
