+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sbtcli-poc",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "A Node CLI for SBT",
   "main": "./lib/sbtcli.js",
   "bin": {
